requirements.txt
langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.5
langchain-google-genai==2.0.5
chromadb==0.5.20
streamlit==1.40.1
boto3==1.35.67
pypdf==5.1.0
python-dotenv==1.0.1
requests==2.32.3
numpy==2.1.3
tqdm==4.67.1
